% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_enrichment.R
\name{barplot_enrichment}
\alias{barplot_enrichment}
\title{Create and save a horizontal barplot of enrichment results}
\usage{
barplot_enrichment(
  comparisons_enrichment,
  comparison,
  subtitle = "",
  first_n = 30,
  min_size_enrichments = 0,
  max_size_enrichments = Inf,
  save_plot = T
)
}
\arguments{
\item{comparisons_enrichment}{A named list of data frames, one per comparison.}

\item{comparison}{A string indicating the name of the comparison (e.g., "SN_C_vs_SN_PD").}

\item{subtitle}{A string to append to the subtitle of the plot (optional).}

\item{first_n}{Integer. The number of top terms to show (default = 30).}

\item{min_size_enrichments}{Integer. Minimum number of genes in background to consider (default = 0).}

\item{max_size_enrichments}{Integer. Maximum number of genes in background to consider (default = Inf).}

\item{save_plot}{Logical. Whether to save the plot as an SVG file (default = TRUE).}
}
\value{
A ggplot object representing the enrichment barplot.
}
\description{
This function generates a horizontal barplot from functional enrichment results
(e.g., GO terms or pathways), showing the proportion of enriched genes (`ratio_in_set`)
per term, colored by FDR. Only terms within the specified gene set size range are shown.
}
\examples{
\dontrun{
  my_enrichment <- enrichment_manova(prot_manova, tax_ID, categories)
  barplot_enrichment(my_enrichment, comparison = "GroupA_vs_GroupB")
}
}
