% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biggest_component.R
\name{biggest_component}
\alias{biggest_component}
\title{Biggest Component of a Graph}
\usage{
biggest_component(graph)
}
\arguments{
\item{graph}{An `igraph` object representing the graph to be analyzed.}
}
\value{
An `igraph` object representing the biggest component of the input graph.
}
\description{
This function checks the number of components in a graph and returns the largest
component if there is more than one.
}
\examples{
library(igraph)
g <- make_ring(10) \%>\%
  add_vertices(5) \%>\%
  add_edges(c(1, 11, 1, 12, 1, 13, 1, 14))
biggest_graph <- biggest_component(g)

}
