% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridging_centrality.R
\name{bridging_centrality}
\alias{bridging_centrality}
\title{Bridging Centrality Calculation}
\usage{
bridging_centrality(
  graph,
  betw = NULL,
  weights_as_distances = NULL,
  range_weights = NULL,
  weighted_degree = T
)
}
\arguments{
\item{graph}{An `igraph` object representing the graph.}

\item{betw}{A precomputed betweenness centrality vector to speed up computation.
If NULL, betweenness will be computed within the function. Default is NULL.}

\item{weights_as_distances}{A vector of weights to be treated as distances on the graph's edges.
If NULL, unweighted bridging centrality will be computed. Default is NULL.}

\item{range_weights}{A numeric range indicating the desired range of weights for the computation.
If NULL, the range of `weights_as_distances` will be used. Ignored if `weighted_degree = FALSE`.
Default is NULL.}

\item{weighted_degree}{A logical indicating whether the degree should be computed as the sum of weights
(when `TRUE`) or the number of edges (when `FALSE`). Default is TRUE.}
}
\value{
A numeric vector of bridging centrality values for each vertex.
}
\description{
This function calculates the bridging centrality of each vertex in a graph. The
bridging centrality combines both the betweenness centrality and the "bridging
coefficient," a measure of how well-connected a node is relative to its neighbors.
The bridging coefficient is calculated following the method described by Hwang et al.
(2006). For weighted graphs, the formula is slightly changed: instead of using
the degree,  For more details, see the reference below.
}
\details{
When `weighted_degree = TRUE`, instead of using the degree (i.e., the number of edges)
as described by Hwang et al, the weighted degree is used, which is calculated
as the sum of the incident edge weights on each vertex.
}
\examples{
library(igraph)
g <- make_ring(10) \%>\%
  add_edges(c(1, 5, 2, 6))
bc <- bridging_centrality(g)

}
\references{
Hwang, W., Cho, Y. R., Zhang, A., & Ramanathan, M. (2006). Bridging Centrality: Identifying Bridging Nodes in Scale-Free Networks. *Technical Report*. Department of Computer Science and Engineering, University at Buffalo, State University of New York. Available at: \url{https://cse.buffalo.edu/tech-reports/2006-05.pdf}.
}
