% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_quantiles.R
\name{centrality_quantiles}
\alias{centrality_quantiles}
\title{centrality_quantiles}
\usage{
centrality_quantiles(graph, fun_list, quantiles = 0, orderBy = NULL, ...)
}
\arguments{
\item{graph}{igraph object}

\item{fun_list}{list of centrality functions to apply.
Use only nodes centralities or only edges centralities.}

\item{quantiles}{threshold of individual centrality quantiles for nodes.
If 0 it will gives all nodes centralities. Default = 0}

\item{orderBy}{integer or name of the centrality corresponding to the
centrality position in 'fun_list' by which the data.frame rows are ordered.
Default as 'rownames'}

\item{...}{evantual further inputs for centrality functions. All functions that
accept a named parameter will accept it if given}
}
\value{
A 'data.frame' with genes as 'rownames', and columns as centrality measures
}
\description{
centrality_quantiles
}
