% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taboo_words.R
\name{check_taboo_words}
\alias{check_taboo_words}
\title{Check for Taboo Words in Character Strings}
\usage{
check_taboo_words(char)
}
\arguments{
\item{char}{A character vector to be checked.}
}
\value{
None. The function is called for its side effect of stopping execution if taboo words are found.
}
\description{
This function checks whether any element of a character vector contains taboo words.
If a taboo word is found, the function throws an error.
}
\details{
The function is case-insensitive and checks if any taboo word is present as a substring within the input characters.
Taboo words include: "fc", "specific", "centrality", "weighted", "number_of_genes", "p_value", "fdr", and "input_genes".
}
\examples{
# Example where no taboo words are present
check_taboo_words(c("gene", "expression"))

# Example where taboo words are present (will stop with an error)
\dontrun{
check_taboo_words(c("p_value", "gene"))
}

}
