% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_centrality_table.R
\name{cn_centrality_table}
\alias{cn_centrality_table}
\title{Extract centralities from critical nodes}
\usage{
cn_centrality_table(centralities, cn_modes)
}
\arguments{
\item{centralities}{centralities output from a `network_analysis` function
(`network_analysis`,`unweighted network_analysis`,`weighted network_analysis`).}

\item{cn_modes}{cn_modes output from a `network_analysis` function
(`network_analysis`,`unweighted network_analysis`,`weighted network_analysis`).}
}
\value{
a list of dataframes with the centralities of the critical nodes
}
\description{
Extract centralities from critical nodes
}
\examples{
\dontrun{
  na <- network_analysis(data_grouped, data_grouped_even_dim, )
  cn_centrality_table(centralities, cn_modes)
}
}
