% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_specific_hubs.R
\name{condition_specific_hubs}
\alias{condition_specific_hubs}
\title{condition_specific_hubs}
\usage{
condition_specific_hubs(hubsList)
}
\arguments{
\item{hubsList}{List of Hubs or bottlenecks data.frames or another output like
the one from 'centrality_quantiles' function}
}
\value{
The list of Hubs or bottlenecks data.frames filtered by specificity,
    e.g. if a Hubs is present on 2 elements of the list, it is removed
}
\description{
condition_specific_hubs
}
