% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ID.R
\name{convert_ID}
\alias{convert_ID}
\title{convert_ID}
\usage{
convert_ID(dataset, gene_column, tax_ID)
}
\arguments{
\item{dataset}{matrix or data.frame}

\item{gene_column}{integer. the position of the gene column}

\item{tax_ID}{integer. Taxonomy ID of the specie}
}
\value{
The dataset with genes converted
}
\description{
convert_ID
}
