% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corr_PPI.model.R
\name{Corr_PPI.model}
\alias{Corr_PPI.model}
\title{Correlation and PPI Network Model Construction}
\usage{
Corr_PPI.model(
  dataset,
  genes_id = NULL,
  Corr = NULL,
  corr.test = "spearman",
  signifCorr = 0.05,
  correctionCorr = "BH",
  interactome.graph = NULL,
  score_threshold = c(escore = 0.15, dscore = 0.35),
  tax_ID = 9606,
  NAasZero = F,
  ZeroasNA = T,
  compute_weights = T
)
}
\arguments{
\item{dataset}{A matrix or data.frame containing the dataset to build network models.}

\item{genes_id}{A character vector of gene identifiers to build the interactome. This is ignored if
\code{interactome.graph} is provided.}

\item{Corr}{A correlation matrix. If provided, the function will use it instead of computing a new one. Default is \code{NULL}.}

\item{corr.test}{A character string specifying the correlation test to use. Options are \code{"spearman"} or \code{"pearson"}.
Default is \code{"spearman"}.}

\item{signifCorr}{A numeric value indicating the significance threshold for correlations. Default is \code{0.05}.}

\item{correctionCorr}{A character string specifying the p-value adjustment method for multiple testing corrections.
Default is \code{"BH"} (Benjamini-Hochberg).}

\item{interactome.graph}{An \code{igraph} object representing the interactome. If \code{NULL}, the interactome is built
using \code{genes_id}. Default is \code{NULL}.}

\item{score_threshold}{A numeric vector specifying thresholds for filtering the interactome.
It should contain two elements: \code{escore} and \code{dscore}. Default is \code{c(escore = 0.15, dscore = 0.35)}.}

\item{tax_ID}{An integer specifying the taxonomy ID of the species under consideration. The default is \code{9606} (human).}

\item{NAasZero}{A logical value. If \code{TRUE}, all \code{NA} values in the dataset will be converted to zeros. Default is \code{FALSE}.}

\item{ZeroasNA}{A logical value. If \code{TRUE}, all zeros in the dataset will be converted to \code{NA}. Default is \code{TRUE}.}

\item{compute_weights}{A logical value. If \code{TRUE}, non-significant correlations will be transformed using a weighting scheme.
Default is \code{TRUE}.}
}
\value{
An \code{igraph} object representing the combined correlation and PPI network model.
}
\description{
This function constructs a network model based on correlations and protein-protein interactions (PPI).
It can compute correlation matrices for the given dataset, integrate them with an existing interactome
network, and optionally transform non-significant correlations into weights.
}
\details{
The function constructs a network model by first creating a correlation matrix from the given dataset.
If \code{interactome.graph} is not provided, the function will build an interactome based on the specified
\code{genes_id} using the STRING database through \code{rbioapi}. The correlation matrix is then integrated
with the interactome to form a combined network model. Optionally, non-significant correlations can be
transformed into weights using the \code{cor2W_transform} function.
}
\examples{
\dontrun{
# Example dataset
dataset <- data.frame(
 Gene1 = c(0.5, NA, 0.3, 0.6, 0.9),
 Gene2 = c(0.4, 0.6, 0, 0.8, 0.4),
 Gene3 = c(NA, 0.1, 0.2, 0.6, 0.9),
 Gene4 = c(0.7, 0.1, 0.2, 0.4, 0.1),
 row.names = c("Sample1", "Sample2", "Sample3", "Sample4", "Sample5")
)

# List of genes
genes_id <- c("Gene1", "Gene2", "Gene3", "Gene4")

# Build the network model
network_model <- Corr_PPI.model(
  dataset = dataset,
  genes_id = genes_id,
  corr.test = "spearman",
  signifCorr = 0.05,
  tax_ID = 9606,
  NAasZero = TRUE,
  ZeroasNA = FALSE
)
}

}
\seealso{
\code{\link[rbioapi]{rba_string_map_ids}}, \code{\link[rbioapi]{rba_string_interactions_network}}, \code{\link[igraph]{graph_from_edgelist}}
}
