% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_nodes.R
\name{critical_nodes}
\alias{critical_nodes}
\title{Identify specific critical nodes in a dataset}
\usage{
critical_nodes(
  x,
  names_of_groups,
  centralities,
  searching_mode,
  ignore_case = T
)
}
\arguments{
\item{x}{A data frame containing the data on groups and centralities.}

\item{names_of_groups}{A vector of group names. Each name should correspond to
one or more columns in the data frame.}

\item{centralities}{A vector of centralities to search for. These should match
the names of the centrality columns in the data frame.}

\item{searching_mode}{A string that specifies the search mode. Options are:
\describe{
  \item{"specific" or "s"}{ - Finds critical nodes that are present in one group but
      absent in all others.}
  \item{"centralityspecific" or "cs"}{ - Finds critical nodes that are specific to
      centralities within one group and absent in the centralities of other groups.}
 }}

\item{ignore_case}{A logical that indicate whether the names_of_groups should be searched with case sensitive.}
}
\value{
A list of data frames, where each data frame contains the critical nodes specific
  to the corresponding group.
}
\description{
This function identifies critical nodes that are specific to certain groups based on
centralities and the selected search mode. It can filter the dataset to return
only those critical nodes that are exclusive to a particular group or to particular centralities.
}
\examples{
# Example data frame
x <- data.frame(groupA_centrality1 = c(1, 0, 1),
                groupA_centrality2 = c(1, 0, 1),
                groupB_centrality1 = c(0, 1, 1),
                groupB_centrality2 = c(0, 1, 0),
                row.names = c("gene1", "gene2", "gene3"))

# Define groups and centralities
names_of_groups <- c("groupA", "groupB")
centralities <- c("centrality1", "centrality2")

# Find group-specific critical nodes
critical_nodes(x, names_of_groups, centralities, "specific")

}
\keyword{bottlenecks}
\keyword{hubs}
