% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_volcano.R
\name{data_volcano}
\alias{data_volcano}
\title{Generate  data for Volcano Plots for Group Comparisons}
\usage{
data_volcano(dataset, manova_pairw_results, significance, gene_column = NULL)
}
\arguments{
\item{dataset}{A data frame where columns represent features (e.g., genes)
and rows represent samples. The last row should contain p-values.}

\item{manova_pairw_results}{Output from `manova_pairwise`}

\item{significance}{A numeric value specifying the significance threshold.}

\item{gene_column}{A numeric or character value indicating the index or the name of the gene column}
}
\value{
A list of transposed data.frame where the last row is the p.value adjusted, useful for volcano plot.
}
\description{
This function generates transposed data.frames volcano plots for pairwise comparisons between groups
in a given dataset.
}
\examples{
set.seed(1)
dataset <- data.frame(GeneName = letters,
                      matrix(runif(26*15), nrow = 26))
colnames(dataset)[-1] <- paste0(rep(c("Group1", "Group2", "Group3"), each = 5), "_",1:15)

mp <- manova_pairwise(dataset,c("Group1", "Group2", "Group3"),gene_column = 1 )
dv <- data_volcano(dataset, mp, 0.95)

vp <- volcano_plot(dv[[1]], 0.95)
}
\seealso{
[volcano_plot()]
}
