% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfList_on_sheet.R
\name{dfList_on_sheet}
\alias{dfList_on_sheet}
\title{Write a List of Data Frames to an Excel Sheet}
\usage{
dfList_on_sheet(
  df_list,
  wb = NULL,
  name_sheet = "Sheet",
  by_col = T,
  rowNames = T,
  n_space = 1
)
}
\arguments{
\item{df_list}{A named or unnamed list of data frames to be written to the Excel sheet.}

\item{wb}{An optional `openxlsx` workbook object. If `NULL`, a new workbook is created.}

\item{name_sheet}{A character string specifying the name of the sheet where the data will be written.
Default is `"Sheet"`.}

\item{by_col}{A logical value indicating whether data frames should be arranged column-wise (`TRUE`)
or row-wise (`FALSE`). Default is `TRUE`.}

\item{rowNames}{A logical value indicating whether row names should be included in the output.
Default is `TRUE`.}

\item{n_space}{An integer specifying the number of empty spaces between tables when arranged
in columns (`by_col = TRUE`) or rows (`by_col = FALSE`). Default is `1`.}
}
\value{
A modified `openxlsx` workbook object (`wb`) with the data written to the specified sheet.
}
\description{
This function writes a list of data frames (`df_list`) to a specified sheet (`name_sheet`)
in an Excel workbook (`wb`). Each data frame is placed either in columns or in rows, depending on `by_col`.
}
\examples{
\dontrun{
library(openxlsx)
df1 <- data.frame(A = 1:5, B = 6:10)
df2 <- data.frame(X = 11:15, Y = 16:20)
df_list <- list(Data1 = df1, Data2 = df2)
wb <- dfList_on_sheet(df_list)
saveWorkbook(wb, "example.xlsx", overwrite = TRUE)
}

}
