% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emPAI.R
\name{emPAI_function}
\alias{emPAI_function}
\title{emPAI}
\usage{
emPAI_function(list.groups_peptides, Pn_ideal)
}
\arguments{
\item{list.groups_peptides}{list of groups of which we want to compute emPAI index, or a dataset Sample x Protein}

\item{Pn_ideal}{vector of ideal peptides}
}
\value{
The emPAI matrix/vector for each protein
}
\description{
emPAI
}
