% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_manova.R
\name{enrichment_manova}
\alias{enrichment_manova}
\title{Perform STRING Enrichment Analysis on manova Results}
\usage{
enrichment_manova(prot_manova, tax_ID, categories, differences = T)
}
\arguments{
\item{prot_manova}{A list where each element contains a vector of protein identifiers (e.g., resulting from manova analysis).}

\item{tax_ID}{A numeric taxonomy ID (e.g., 9606 for Homo sapiens) to be used in the STRING enrichment analysis.}

\item{categories}{A character vector specifying the categories of enrichment terms to filter (e.g., "GO:BP", "KEGG").}

\item{differences}{Logical, if TRUE (default), computes the difference in enrichment terms between consecutive pairs of groups.}
}
\value{
If `differences = TRUE`, returns a list containing:
\item{enr_manova}{The original enrichment results for each group.}
\item{enr_manova.diff}{The differential enrichment terms between consecutive groups.}
If `differences = FALSE`, returns a list of enrichment results for each group.
}
\description{
This function performs STRING enrichment analysis on a list of proteins (manova results), filters the results based on specified categories, and can optionally compute the differences between enrichment terms in consecutive pairs of groups.
}
\details{
This function utilizes the STRING enrichment API via the `rbioapi::rba_string_enrichment` function to perform enrichment analysis for each group of proteins. It filters the enrichment results based on the specified categories and can optionally compute the difference in terms between consecutive pairs of groups.
}
\examples{
\dontrun{
# Example protein manova results
prot_manova <- list(c("P12345", "P67890"), c("P54321", "P09876"))
tax_ID <- 9606  # Homo sapiens
categories <- c("GO:BP", "KEGG")
enrichment_manova(prot_manova, tax_ID, categories)
}
}
