% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.R
\name{FC}
\alias{FC}
\title{Fold Change (FC) Index Calculation for Pairwise Group Comparisons}
\usage{
FC(groups_list)
}
\arguments{
\item{groups_list}{A list where each element represents a group of observations (e.g., numeric vectors or matrices of data).}
}
\value{
A matrix where each column corresponds to the log2 Fold Change (FC) for pairwise comparisons between groups.
The column names follow the pattern "Group1_vs_Group2" and "Group2_vs_Group1" for each pairwise comparison.
}
\description{
Computes the Fold Change (FC) for pairwise comparisons between groups of observations.
}
\examples{
group1 <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2)
group2 <- matrix(c(3, 2, 1, 6, 5, 4), nrow = 2)
group3 <- matrix(c(5, 4, 3, 2, 1, 0), nrow = 2)
groups_list <- list(Group1 = group1, Group2 = group2, Group3 = group3)
groups_list <- lapply(groups_list, function(x){
    rownames(x) <- c("gene1","gene2")
    colnames(x) <- paste0("sample_",1:3)
    return(x)
})
FC(groups_list)

}
