% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_quantiles.R
\name{filter_quantiles}
\alias{filter_quantiles}
\title{filter_quantiles}
\usage{
filter_quantiles(centralities, quantiles, orderBy = NULL)
}
\arguments{
\item{centralities}{data.frame of centralities to filter}

\item{quantiles}{quantile to filter}

\item{orderBy}{integer or name of the centrality corresponding to the
centrality position in 'colnames(centralities)' by which the data.frame rows are ordered.
Default as 'rownames'}
}
\value{
A data.frame filtered
}
\description{
filter_quantiles
}
