% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten2CorrPPI.R
\name{flatten2CorrPPI}
\alias{flatten2CorrPPI}
\title{flatten2CorrPPI}
\usage{
flatten2CorrPPI(Corr)
}
\arguments{
\item{Corr}{output of Hmisc::rcorr()}
}
\value{
A flatten correlation data.frame, used for 'Corr_PPI.model()' function
}
\description{
flatten2CorrPPI
}
