% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_listing.R
\name{group_listing}
\alias{group_listing}
\title{group.listing}
\usage{
group_listing(
  dataset.t,
  names_of_groups,
  ignoreCase = T,
  pos.vectors_groups = NULL
)
}
\arguments{
\item{dataset.t}{Dataset Samples x Proteins}

\item{names_of_groups}{ignored if pos.vectors_groups is not NULL. Names of condition specified within rownames(dataset)}

\item{ignoreCase}{ignored if 'names_of_groups' is not used. If FALSE the names in 'names_of_groups' are case sensitives. default T.}

\item{pos.vectors_groups}{list with in each element a vector with the row indexes}
}
\value{
A list with each specified group
}
\description{
group.listing
}
