% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtest.R
\name{Gtest}
\alias{Gtest}
\title{G Test for Pairwise Group Comparisons}
\usage{
Gtest(list.groups)
}
\arguments{
\item{list.groups}{A list where each element represents a group of observations (e.g., numeric vectors or matrices of data).}
}
\value{
A list of G values, one for each pairwise comparison between groups.
}
\description{
Computes the G statistic for pairwise comparisons between groups. The G test compares two groups of data and evaluates the differences between them using a log-likelihood ratio.
}
\details{
The G test statistic is computed as:
\deqn{2 * g1 * log(g1 / ((g1 + g2) / 2)) + 2 * g2 * log(g2 / ((g1 + g2) / 2))} where \code{g1} and \code{g2} represent two groups being compared. The sign of the difference is retained to indicate the direction of the comparison.
}
\examples{
group1 <- c(1, 2, 3, 4, 5)
group2 <- c(2, 3, 4, 5, 6)
group3 <- c(3, 4, 5, 6, 7)
list.groups <- list(Group1 = group1, Group2 = group2, Group3 = group3)
Gtest(list.groups)

}
