% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\title{Linear Discriminant Analysis}
\usage{
LDA(
  dataset,
  names_of_groups,
  gene_column,
  ignoreCase = T,
  significance.LDA = 0.05,
  correction.LDA = "BH",
  pos.vectors_groups = NULL,
  boundFC = c(-2, 2),
  MDS = T
)
}
\arguments{
\item{dataset}{The dataset to analyze}

\item{names_of_groups}{names of groups in colnames(dataset), ignored if 'pos.vector_groups' is not NULL}

\item{gene_column}{position of gene column. If 0 no genes column is present}

\item{ignoreCase}{if TRUE the 'names_of_groups' search is not case sensitive. Default = T}

\item{significance.LDA}{significance threshold for features pvalues. Default = 0.05}

\item{correction.LDA}{correction method to adjusting pvalues. Default = "BH"}

\item{pos.vectors_groups}{List containing column indexes for each group to compare}

\item{boundFC}{Numeric vector. Interval out of which labels are highlithed in red and labels are shown}

\item{MDS}{If TRUE computes random forest classification and Multi Dimensional Scaling. Default = T}
}
\value{
A list with linear discriminant analysis information
}
\description{
Linear Discriminant Analysis
}
