% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova.R
\name{manova}
\alias{manova}
\title{Multivariate Analysis of Variance (manova)}
\usage{
manova(
  dataset,
  names_of_groups,
  gene_column,
  ignore_case = T,
  correction_manova = "BH",
  pos_vectors_groups = NULL,
  fc_bounds = c(-0, 0)
)
}
\arguments{
\item{dataset}{The dataset to analyze, where rows represent features (e.g., genes, proteins) and columns represent samples.}

\item{names_of_groups}{A character vector specifying the names of the groups present in the column names of the dataset. Ignored if `pos_vectors_groups` is provided.}

\item{gene_column}{The position of the column containing gene identifiers. If set to 0, no gene column is assumed.}

\item{ignore_case}{Logical, if TRUE (default), the matching of `names_of_groups` is case insensitive.}

\item{correction_manova}{A character string specifying the p-value correction method. Default is "BH" (Benjamini-Hochberg). See `p.adjust.methods` for more options.}

\item{pos_vectors_groups}{A list where each element contains column indices for a group of samples. If provided, `names_of_groups` is ignored.}

\item{fc_bounds}{A numeric vector specifying the bounds for highlighting fold changes in Volcano plots. Default is c(0, 0).}
}
\value{
A data.frame with p-values, adjusted p-values, Fold changes and DAves for the features.
}
\description{
This function performs  Multivariate Analysis of Variance on a dataset, identifying significant features across multiple groups.
}
\examples{
dataset <- matrix(rnorm(100), nrow = 20)
colnames(dataset) <- c("control_1", "control_2", "treatment_1", "treatment_2", "control_3")
rownames(dataset) <- paste("gene", 1:20, sep = "_")
manova(dataset, names_of_groups = c("control", "treatment"), gene_column = 0)

}
