% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova_pairwise.R
\name{manova_pairwise}
\alias{manova_pairwise}
\title{Pairwise Multivariate Analysis of Variance (manova)}
\usage{
manova_pairwise(dataset, names_of_groups, gene_column, ...)
}
\arguments{
\item{dataset}{A data frame containing the data for the analysis.}

\item{names_of_groups}{A character vector specifying the names of the groups to be compared.}

\item{gene_column}{The name or index of the column in the dataset that contains the gene identifiers.}

\item{...}{Additional arguments to be passed to the `manova` function.}
}
\value{
A named list of manova results for each pairwise comparison, with names indicating the groups compared.
}
\description{
This function performs pairwise Multivariate Analysis of Variance (manova) for all possible combinations of the specified groups in the dataset. It returns a list of manova results for each pairwise comparison.
}
\examples{
# Example dataset
set.seed(123)
data <- data.frame(
  gene = paste0("gene_", 1:10),
    matrix(rnorm(90), ncol = 9)
    )
colnames(data)[2:10] <- paste0("group", rep(1:3, each = 3),
                               "_", rep(1:3, times = 3))
# Perform pairwise manova
manova_results <- manova_pairwise(data,
                            names_of_groups = c("Group1", "Group2", "Group3"),
                            gene_column = 1)
print(manova_results)
}
