% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_enrichment.R
\name{merged_enrichment}
\alias{merged_enrichment}
\title{Merges enrichment data from a list of enrichment results of up and down regulated genes.}
\usage{
merged_enrichment(enr_list)
}
\arguments{
\item{enr_list}{A named list of data frames containing enrichment results.}
}
\value{
A named list of merged enrichment data frames.
}
\description{
Merges enrichment data from a list of enrichment results of up and down regulated genes.
}
\examples{
\dontrun{
merged_enrichment(enr_list)
}
}
