% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Data normalization}
\usage{
normalization(dataset, norm_type)
}
\arguments{
\item{dataset}{A data frame where rows represent proteins and columns represent samples.}

\item{norm_type}{A character string specifying the type of normalization. Choose from:
\describe{
  \item{ln}{Natural logarithm transformation}
  \item{Znorm}{Z-score normalization}
  \item{MinMax}{Min-Max scaling}
  \item{Robust}{Robust scaling using median and IQR}
  \item{UnitVector}{Unit vector scaling}
  \item{TotSigNorm}{Total signal normalization}
  \item{MaxSigNorm}{Maximum signal normalization}
  \item{RowSigmaNorm}{Row sigma normalization}
}}
}
\value{
A data frame with the normalized data.
}
\description{
This function normalizes a dataset (Proteins x Samples) using various normalization techniques. It supports different methods for scaling and transforming the data, including natural logarithm transformation, Z-score normalization, Min-Max scaling, and more.
}
\examples{
# Example dataset
data <- data.frame(
  Sample1 = c(10, 20, 30),
  Sample2 = c(5, 15, 25),
  Sample3 = c(2, 8, 18)
)

# Normalize using different methods
normalization(data, norm_type = "ln")
normalization(data, norm_type = "Znorm")
normalization(data, norm_type = "MinMax")
normalization(data, norm_type = "Robust")
normalization(data, norm_type = "UnitVector")
normalization(data, norm_type = "TotSigNorm")
normalization(data, norm_type = "MaxSigNorm")
normalization(data, norm_type = "RowSigmaNorm")

}
