% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSAF.R
\name{NSAF}
\alias{NSAF}
\title{NSAF}
\usage{
NSAF(list.groups, length_proteins)
}
\arguments{
\item{list.groups}{list of groups of which we want to compute NSAF index, or a dataset Samples x Proteins}

\item{length_proteins}{length of proteins, ordered according to the datasets}
}
\value{
A matrix/vector of nsaf indexes for each protein
}
\description{
NSAF
}
