% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline for Proteomics Data Analysis}
\usage{
pipeline(
  dataset,
  names_of_groups,
  gene_column = 1,
  norm_type = "TotSigNorm",
  quantile_critical_nodes = 0.75,
  significance_manova = 0.05,
  fun_list = c(Betweenness = igraph::betweenness, Centroids = centroids, Bridging =
    bridging_centrality),
  g_interactome = NULL,
  categories = c("Component", "Function", "Process", "RCTM", "WikiPathway"),
  tax_ID = 9606,
  save_results_as = paste0("PiProteline_report_", paste0(names_of_groups, collapse =
    "_")),
  ...
)
}
\arguments{
\item{dataset}{A data frame where rows represent proteins and columns represent samples.}

\item{names_of_groups}{A character vector specifying the groups within `colnames(dataset)` for grouping the data.}

\item{gene_column}{The name or index of the column containing gene identifiers. Default is 1.}

\item{norm_type}{The type of normalization to apply. Default is "TotSigNorm".}

\item{quantile_critical_nodes}{A numeric value specifying the quantile threshold for identifying critical nodes. Default is 0.75.}

\item{significance_manova}{A numeric argument indicating the significance threshold for Multivariate Analysis of Variance}

\item{fun_list}{A named list of functions for calculating centralities. Default is a set of common centrality measures.}

\item{g_interactome}{An `igraph` object representing the interactome of the species, used for mapping protein interactions.}

\item{categories}{A character vector specifying enrichment categories for functional analysis. Default is `c("Component", "Function", "Process", "RCTM", "WikiPathway")`.}

\item{tax_ID}{An integer representing the taxonomy ID for the organism (e.g., 9606 for human).}

\item{save_results_as}{Directory name where results will be saved. Defaults to "PiProteline_report".}

\item{...}{Additional arguments passed to other functions within the pipeline.}
}
\value{
A list containing results from different steps of the analysis, including:
\describe{
  \item{descriptiveStatistics}{Data grouped, column and row statistics, correlation between groups, and indices like NSAF and emPAI.}
  \item{quantitativeAnalysis}{Results of manova analysis, DAve and DCI indices, and fold change calculations.}
  \item{networkAnalysis}{Unweighted and weighted PPI analysis, centralities, critical nodes, and violin plots.}
  \item{functionalAnalysis}{Results of the functional analysis, including enrichment and intersecting enrichment terms.}
}
}
\description{
This function implements a multi-step pipeline for analyzing a dataset, which includes data preprocessing, normalization, quantitative analysis, network topology analysis, and visualization using violin plots.
}
\examples{
\dontrun{
# Example of how to call the pipeline function
dataset <- matrix(rnorm(100), nrow = 10)
colnames(dataset) <- c("control_1", "control_2", "treatment_1", "treatment_2",
                       "control_3", "treatment_3", "control_4", "control_5",
                       "treatment_4", "treatment_5")
result <- pipeline(dataset, names_of_groups = c("control", "treatment"), tax_ID = 9606)
}

}
