% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_analysis.R
\name{quantitative_analysis}
\alias{quantitative_analysis}
\title{Quantitative Analysis of Dataset}
\usage{
quantitative_analysis(
  dataset,
  names_of_groups,
  gene_column,
  data.grouped.full = NULL,
  ...
)
}
\arguments{
\item{dataset}{A data frame containing the data to analyze.}

\item{names_of_groups}{A character vector specifying the names of the groups for analysis.}

\item{gene_column}{An integer or character specifying the column in the dataset containing gene names.}

\item{data.grouped.full}{List of data frames with the same number of rows.}

\item{...}{Additional arguments passed to the underlying functions for LDA, DAve, DCI, and fold change calculations.}
}
\value{
A list containing the results of the quantitative analysis:
\describe{
  \item{LDA_results}{The results from the linear discriminant analysis (LDA).}
  \item{LDA_pairw.results}{The results from pairwise LDA, if applicable.}
  \item{DAve_index}{The DAve index calculated from the data.}
  \item{DCI_index}{The DCI index calculated from the data.}
  \item{Fold_Change}{The fold change analysis results.}
}
}
\description{
This function performs quantitative analysis on a dataset, including linear discriminant analysis (LDA), pairwise LDA (if more than two groups are provided), and calculation of additional indices such as DAve, DCI, and fold change.
}
\examples{
# Example dataset
data <- data.frame(
matrix(runif(1500),ncol = 15,
       dimnames = list(paste0("Gene_",1:100),
                       paste0("Group",rep(c(1,2,3),each = 5),"_",c(1:5))))
)
data$Gene <- paste0("Gene_",1:100)

# Perform quantitative analysis
quantitative_analysis(data, names_of_groups = c("Group1", "Group2", "Group3"),
                      gene_column = "Gene")

}
