% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{remove_duplicates}
\usage{
remove_duplicates(dataset, genes, FUN = max, NAasZero = T, ZeroasNA = F)
}
\arguments{
\item{dataset}{The dataset in which we want to remove duplicates}

\item{genes}{name of the column containing the Gene names, or vector containing
as elements as the number of dataset rows}

\item{FUN}{the function applied to duplicated values, default max}

\item{NAasZero}{if TRUE all NAs will be converted in zeros, default T
if both NAasZero and ZeroasNA are TRUE, it will give an error}

\item{ZeroasNA}{if TRUE all zeros will be converted in NAs, default F}
}
\value{
A data.frame with no more duplicated Genes, with a new gene column called GeneName
}
\description{
remove_duplicates
}
