% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_list.R
\name{reverse_list}
\alias{reverse_list}
\title{reverse_list}
\usage{
reverse_list(originalList)
}
\arguments{
\item{originalList}{list of lists composed the same between each others}
}
\value{
A list with elements switched of hierarchy
}
\description{
reverse_list
}
