% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_images.R
\name{save_images}
\alias{save_images}
\title{Save MDS and Volcano plots from quantitative analysis}
\usage{
save_images(quantitative_analysis_results, device = "svg")
}
\arguments{
\item{quantitative_analysis_results}{A list containing MDS and volcano plots.}

\item{device}{The file format for saving plots (default: "svg").}
}
\description{
Save MDS and Volcano plots from quantitative analysis
}
\examples{
\dontrun{
save_images(quantitative_analysis_results)
}
}
