% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Save Results to Excel Files}
\usage{
save_results(
  quantitative_analysis_results = NULL,
  network_analysis_results = NULL,
  functional_analysis_results = NULL,
  pipeline_results = NULL,
  save_results_as = "PiProteline_report"
)
}
\arguments{
\item{quantitative_analysis_results}{Output of the `quantitative_analysis` function.}

\item{network_analysis_results}{Output of network analysis.}

\item{functional_analysis_results}{Output of functional analysis.}

\item{pipeline_results}{Output of the complete analysis pipeline. If provided, individual results (`quantitative_analysis_results`, `network_analysis_results`, `functional_analysis_results`) are ignored.}

\item{save_results_as}{Directory name where results will be saved. Defaults to "PiProteline_report".}
}
\value{
Saves Excel files in the specified directory.
}
\description{
This function saves the results PiProteline analyses to Excel files, organizing data into different sheets.
}
\examples{
\dontrun{
save_results(quantitative_analysis_results, network_analysis_results,
             functional_analysis_results, save_results_as = "my_results")
}
}
