% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_profile_enrichment.R
\name{single_profile_enrichment}
\alias{single_profile_enrichment}
\title{Single Profile Enrichment Analysis}
\usage{
single_profile_enrichment(
  dataset,
  names_of_groups,
  gene_column = 1,
  tax_ID,
  categories = c("Process", "Function", "Component", "RCTM", "WikiPathways"),
  parallel = T,
  num_cores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{dataset}{A data frame where rows represent genes and columns contain
information across groups. Non-numeric columns are ignored, and all values
greater than 0 are considered as present.}

\item{names_of_groups}{A character vector specifying the names of the groups
in `dataset` for profile enrichment.}

\item{gene_column}{An integer or character specifying the column in `dataset`
that contains gene identifiers. Default is the first column.}

\item{tax_ID}{Numeric species taxonomic ID used in enrichment analysis.}

\item{categories}{A character vector of categories for enrichment filtering.
Default categories include "Process", "Function", "Component", "RCTM", and "WikiPathways".}

\item{parallel}{Logical; if TRUE, the function will execute in parallel mode.
Default is TRUE.}

\item{num_cores}{Integer specifying the number of cores to use for parallel}

\item{...}{Other arguments to pass on internal functions
processing. Default is one less than the total available cores.}
}
\value{
A data frame containing enrichment results per category and term, with
additional columns for computed metrics such as means, DAve, and manova-adjusted
p-values.
}
\description{
This function performs a profile enrichment analysis on a specified dataset by
identifying enriched terms for each group of proteins. The function supports both
parallel and sequential computation modes.
}
\details{
This function first identifies proteins in each group and applies
enrichment analysis. Enrichment results are then filtered by selected categories and
metrics are calculated.
}
\examples{
\dontrun{
dataset <- data.frame(geneID = c("gene1", "gene2", "gene3"),
                      group1 = c(1, 0, 1),
                      group2 = c(0, 1, 0))
single_profile_enrichment(dataset, names_of_groups = c("group1", "group2"),
                          tax_ID = 9606)
}

}
