% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_CriticalNodes.R
\name{summary_CriticalNodes}
\alias{summary_CriticalNodes}
\title{Summary of Critical Nodes}
\usage{
summary_CriticalNodes(CN.output, paste_colnames = "")
}
\arguments{
\item{CN.output}{A list containing the critical nodes output from the network analysis. The list is expected to have searching modes as top-level keys, followed by critical node types ("Hubs" or "Bottlenecks") and group names.}

\item{paste_colnames}{A character string to append to the column names of the resulting matrix. Default is an empty string.}
}
\value{
A list containing:
\describe{
  \item{prot_Netw}{A list where each element contains the critical nodes for a specific combination of searching mode, critical node type, and group.}
  \item{CN.M}{A data frame where rows represent unique critical nodes and columns indicate the presence of each node across various combinations of searching mode, critical node type, and group, with a final column for gene names.}
}
}
\description{
This function generates a summary of critical nodes (Hubs and Bottlenecks) identified in network analysis results for multiple groups. It creates a matrix indicating the presence of each critical node across different groups and searching modes.
}
\examples{
\dontrun{
# Example usage
CN.output <- list(
  SearchMode1 = list(
    Hubs = list(Group1 = data.frame(Gene1 = c(1), Gene2 = c(1))),
    Bottlenecks = list(Group1 = data.frame(Gene3 = c(1)), Group2 = data.frame(Gene4 = c(1)))
  ),
  SearchMode2 = list(
    Hubs = list(Group1 = data.frame(Gene2 = c(1)), Group2 = data.frame(Gene1 = c(1))),
    Bottlenecks = list(Group1 = data.frame(Gene4 = c(1)))
  )
)
names_of_groups <- c("Group1", "Group2")
summary_CriticalNodes(CN.output, paste_colnames = "_CN")
}
}
