% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_LDA.R
\name{summary_LDA}
\alias{summary_LDA}
\title{Summary of LDA Analysis Results}
\usage{
summary_LDA(volcano_plots)
}
\arguments{
\item{volcano_plots}{A list containing LDA results for pairwise comparisons, where each comparison includes a list of volcano plots with up-regulated and down-regulated genes.}
}
\value{
A list containing:
\describe{
  \item{prot_LDA}{A list of up-regulated and down-regulated genes for each comparison.}
  \item{gene_matrix}{A data frame where rows represent unique genes and columns indicate the presence of each gene in up-regulated or down-regulated lists for each comparison, with a final column for gene names.}
}
}
\description{
This function generates a summary of linear discriminant analysis (LDA) results for pairwise comparisons. It extracts up-regulated and down-regulated genes from the LDA results and constructs a gene matrix indicating the presence of each gene across different comparisons.
}
\examples{
\dontrun{
# Example usage
volcano_plots <- list(
  Comparison1 = list(
    VolcanoPlots = list(
      list(list(Genes = list(UP = c("Gene1", "Gene2"), DOWN = c("Gene3"))))
    )
  ),
  Comparison2 = list(
    VolcanoPlots = list(
      list(list(Genes = list(UP = c("Gene3", "Gene4"), DOWN = c("Gene1"))))
    )
  )
)
summary_LDA(volcano_plots)
}
}
