% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{transpose}
\usage{
transpose(dataset, gene_column = 1)
}
\arguments{
\item{dataset}{The dataset to transpose}

\item{gene_column}{names or indexes of the columns that contain characters,
if 0 rownames will be used as colnames. Default = 1}
}
\value{
A data.frame transposed
}
\description{
transpose
}
