% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unweighted_network_analysis.R
\name{unweighted_network_analysis}
\alias{unweighted_network_analysis}
\title{Unweighted Network Analysis}
\usage{
unweighted_network_analysis(
  data.grouped = NULL,
  names_of_groups,
  data.unique = NULL,
  fun_list,
  g.interactome,
  quantile_critical_nodes,
  ...
)
}
\arguments{
\item{data.grouped}{A list of data frames representing grouped data. If `NULL`, data is grouped using `data.unique` and the specified parameters.}

\item{names_of_groups}{A character vector specifying the names of the groups for analysis.}

\item{data.unique}{A data frame with unique data used for grouping, if `data.grouped` is `NULL`. Default is `NULL`.}

\item{fun_list}{A list of functions for calculating network centralities.}

\item{g.interactome}{A igraph object representing the interactome.}

\item{quantile_critical_nodes}{A numeric value specifying the quantile threshold for identifying critical nodes.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A list containing:
\describe{
  \item{Centralities}{Centrality measures for unweighted networks.}
  \item{CriticalNodes}{Lists of critical nodes (hubs and bottlenecks) in unweighted networks.}
}
}
\description{
This function performs network analysis on unweighted networks derived from grouped biological data. It calculates centralities, identifies hubs and bottlenecks, and analyzes centrality over quantiles.
}
\examples{
library(igraph)
# Example data
filteredInteractome <- filter_interactome(interactome_hs,
                                          scores_threshold = c("experimental" = 0.15,
                                                               "database"    = 0.35))
g.interactome <- graph_from_edgelist(as.matrix(filteredInteractome[,3:4]),
                                     directed = FALSE)
set.seed(1)
data.grouped <- list(Group1 = matrix(runif(2000), nrow = 200),
                     Group2 = matrix(runif(2000), nrow = 200))
data.grouped <- lapply(data.grouped, function(x) {
  rownames(x) <- sample(names(V(g.interactome)), nrow(x))
  return(x)
})
colnames(data.grouped$Group1) <- paste0("Group1_",1:10)
colnames(data.grouped$Group2) <- paste0("Group2_",1:10)
result <- unweighted_network_analysis(data.grouped = data.grouped,
                                      names_of_groups = c("Group1", "Group2"),
                                      fun_list = list(Degree = igraph::degree,
                                                     Betweenness = igraph::betweenness),
                                     g.interactome = g.interactome,
                                     quantile_critical_nodes = 0.8)


}
