% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wDegree.R
\name{wDegree}
\alias{wDegree}
\title{Calculate Degree or Weighted Degree of a Graph}
\usage{
wDegree(graph, weights_wd = NULL, ...)
}
\arguments{
\item{graph}{An `igraph` object representing the graph.}

\item{weights_wd}{A character string specifying the name of the edge attribute to use for weighted degree calculation.
If `NULL`, unweighted degree is calculated. Default is `NULL`.}

\item{...}{Additional arguments to be passed to `igraph::degree` if using unweighted calculation.}
}
\value{
A numeric vector containing the degree or weighted degree for each node in the graph.
}
\description{
This function calculates the degree of nodes in a graph, either using standard unweighted degree calculation
 or weighted degree based on a specified edge attribute, after taken their absolute value.
}
\examples{
library(igraph)
# Create a simple graph
g <- make_ring(10)

# Calculate unweighted degree
unweighted_deg <- wDegree(g)

# Add weights to the edges
E(g)$weight <- runif(ecount(g))

# Calculate weighted degree
weighted_deg <- wDegree(g, weights_wd = "weight")

# Print the results
print(unweighted_deg)
print(weighted_deg)
}
