% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_as_distances.R
\name{weights_as_distances}
\alias{weights_as_distances}
\title{weights_as_distances}
\usage{
weights_as_distances(
  originalWeights = NULL,
  graph = NULL,
  range.weights = NULL
)
}
\arguments{
\item{originalWeights}{A vector with the node weights. If NULL it will be
used the first edge.attributes of 'graph'. Default = NULL}

\item{graph}{igraph object. Not necessary if 'originalWeights' is given. Default = NULL}

\item{range.weights}{The range the should have the weights. If NULL it will
be used range(originalWeights) . Default = NULL}
}
\value{
A vector containing the weights
}
\description{
weights_as_distances
}
