% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_after_space.R
\name{wrap_after_space}
\alias{wrap_after_space}
\title{wrap text after a space and number of characters}
\usage{
wrap_after_space(text, n = 20)
}
\arguments{
\item{text}{a character vector to wrap}

\item{n}{an integer specifying the number of characters after which to wrap the text}
}
\value{
a character vector with wrapped text
}
\description{
wrap text after a space and number of characters
}
