% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biggest_component.R
\name{biggest_component}
\alias{biggest_component}
\title{biggest_component}
\usage{
biggest_component(graph)
}
\arguments{
\item{graph}{check the number graph components and take the biggest one if there
are more than one}
}
\value{
igraph object, the biggest component
}
\description{
biggest_component
}
