% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridging_centrality.R
\name{bridging_centrality}
\alias{bridging_centrality}
\title{bridging_centrality}
\usage{
bridging_centrality(
  graph,
  Betw = NULL,
  weights_as_distances = NULL,
  range_weights = NULL,
  degree_weighted = T
)
}
\arguments{
\item{graph}{igraph object}

\item{Betw}{if already computed it can be given to speed up the computation.
Default = NULL}

\item{weights_as_distances}{weights (as distances) of all vertices.
If NULL it will be computed the unweighted bridging. Default = NULL}

\item{range_weights}{The range that should have the weights, used to compute
the weighted degree. If NULL the range is assumed to be range(weights_as_distances).
Ignored if weighted degree is FALSE. Default = NULL}

\item{degree_weighted}{If TRUE the degree will be computed as the sum of the weights
(not as distances), instead of the number of the edges}
}
\value{
A bridging centrality vector
}
\description{
bridging_centrality
}
