% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_mean.R
\name{centrality_mean}
\alias{centrality_mean}
\title{centrality_mean}
\usage{
centrality_mean(graph, fun_list, orderBy = NULL, ...)
}
\arguments{
\item{graph}{igraph object}

\item{fun_list}{list of centrality functions to apply.
Use only nodes centralities or only edges centralities.}

\item{orderBy}{integer or name of the centrality corresponding to the
centrality position in 'fun_list' by which the data.frame rows are ordered.
Default as 'rownames'}

\item{...}{evantual further inputs for centrality functions. All functions that
accept a named parameter will accept it if given}
}
\value{
A 'data.frame' with genes as 'rownames', and columns as centrality measures
}
\description{
centrality_mean
}
