% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality_quantiles.R
\name{centrality_quantiles}
\alias{centrality_quantiles}
\title{Centrality Quantiles Filtering}
\usage{
centrality_quantiles(graph, fun_list, quantiles = 0, orderBy = NULL, ...)
}
\arguments{
\item{graph}{An `igraph` object representing the graph.}

\item{fun_list}{A list of centrality functions to apply. The functions should either
return node centralities or edge centralities, but not a mix.}

\item{quantiles}{A threshold value representing the quantile of individual centrality
measures for nodes. Nodes with a centrality score above the given quantile will
be included. Default is 0 (includes all nodes).}

\item{orderBy}{An integer or the name of the centrality by which to order the
resulting data frame rows. If NULL, the data frame will be ordered by row names.
Default is NULL.}

\item{...}{Additional arguments to pass to the centrality functions. These will be
filtered to match the parameters accepted by each function.}
}
\value{
A data frame with nodes as row names and centrality measures as columns.
    Only nodes with a centrality score above the given quantile for each measure
    will be included.
}
\description{
This function applies multiple centrality measures to a graph and filters nodes
based on whether their centrality score is above the quantile threshold specified.
It returns a data frame containing only the common nodes that meet this condition across
all centrality measures.
}
\details{
You can choose to order the resulting data frame by a specific centrality measure
using the `orderBy` argument.
}
\examples{
library(igraph)
g <- make_ring(10)
centralities <- centrality_quantiles(g, fun_list = list(degree, closeness), quantiles = 0.5)
print(centralities)

}
