% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids.R
\name{centroids}
\alias{centroids}
\title{centroids}
\usage{
centroids(
  graph,
  weights_as_distances = NULL,
  parallel = T,
  num_cores = round(3 * detectCores()/4)
)
}
\arguments{
\item{graph}{igraph object}

\item{weights_as_distances}{weights (as distances) used to compute the shortest paths.
If NULL centroids will be computed unweighted. Default = NULL}

\item{parallel}{if TRUE it will be used 'num_cores' to parallel the process.
Default = T}

\item{num_cores}{number of cores to be used, ignored if parallel = FALSE.
Default = round(3*detectCores()/4)}
}
\value{
centroids vector for all nodes
}
\description{
centroids
}
