% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2W_transform.R
\name{cor2W_transform}
\alias{cor2W_transform}
\title{cor2W_transform}
\usage{
cor2W_transform(flattCorr, signifCorr)
}
\arguments{
\item{flattCorr}{correlation martix flatten as data.frame}

\item{signifCorr}{significance threshold}
}
\value{
correlation matrix flatten as data.frame with weights column
}
\description{
cor2W_transform
}
