% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corr_PPI.model.R
\name{Corr_PPI.model}
\alias{Corr_PPI.model}
\title{Title}
\usage{
Corr_PPI.model(
  list.groups,
  genes_id,
  list.Corr = NULL,
  corr.test = "spearman",
  signifCorr = 0.05,
  correctionCorr = "BH",
  interactome = NULL,
  tax_ID = 9606,
  NAasZero = F,
  ZeroasNA = T,
  compute_weights = T
)
}
\arguments{
\item{list.groups}{List of groups to build network models}

\item{genes_id}{vector of genes to build interactome, ignored if 'interactome' is given}

\item{list.Corr}{list of flatten correlations to speed up the computation.
See 'flatten2CorrPPI' function. Default = NULL}

\item{corr.test}{"spearman" or "pearson". Default = "spearman"}

\item{signifCorr}{significance threshold. Default = 0.05}

\item{correctionCorr}{correction method for correlation pvalues. Default = "BH"}

\item{interactome}{matrix or data.frame interactome. If NULL is built on 'genes_id'. default = NULL}

\item{tax_ID}{taxonomy ID of the spices into consideration. default human = 9606}

\item{NAasZero}{if TRUE all NAs will be converted in zeros, default F
if both NAasZero and ZeroasNA are TRUE, it will give an error}

\item{ZeroasNA}{if TRUE all zeros will be converted in NAs, default T}

\item{compute_weights}{if TRUE non significant correlations will be transformed.
Default = T}
}
\value{
A list with PPI/Co-expression network model for each condition
}
\description{
Title
}
