% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_interactome.R
\name{filter_interactome}
\alias{filter_interactome}
\title{filter_interactome}
\usage{
filter_interactome(
  interactome,
  scores_threshold = c(escore = 150, dscore = 300)
)
}
\arguments{
\item{interactome}{data.frame or matrix interactome}

\item{scores_threshold}{a numeric vector named with the types of scores to filter.
Default =  c("escore" = 0.15, "dscore" = 0.3)}
}
\value{
A data.frame interactome filtered
}
\description{
filter_interactome
}
