% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenCorrMatrix.R
\name{flattenCorrMatrix}
\alias{flattenCorrMatrix}
\title{flattenCorrMatrix}
\usage{
flattenCorrMatrix(cormat, pmat, nmat)
}
\arguments{
\item{cormat}{matrix of correlation values}

\item{pmat}{matrix of correlation pvalues}

\item{nmat}{matrix of observation pairs}
}
\value{
A data.frame that flattens the correlation matrices
}
\description{
flattenCorrMatrix
}
