% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{functional_analysis}
\alias{functional_analysis}
\title{Functional Analysis for Network and manova Results}
\usage{
functional_analysis(
  dataset,
  manova_pairwise_results,
  unweighted_CN,
  weighted_CN,
  names_of_groups,
  tax_ID,
  categories,
  ...
)
}
\arguments{
\item{dataset}{A data.frame containing protein samples, with column named after names of groups}

\item{manova_pairwise_results}{The results from pairwise manova analysis.}

\item{unweighted_CN}{The list of critical nodes identified in unweighted networks.}

\item{weighted_CN}{The list of critical nodes identified in weighted networks.}

\item{names_of_groups}{A character vector specifying the names of the groups for analysis.}

\item{tax_ID}{An integer representing the taxonomy ID for the organism (e.g., 9606 for human).}

\item{categories}{A character vector specifying the enrichment categories to filter (e.g., "GO", "KEGG").}

\item{...}{Additional arguments passed to \link{single_profile_enrichment} function.}
}
\value{
A list containing:
\describe{
  \item{summaryTable}{A list containing summary data for manova results, unweighted, and weighted networks.}
  \item{enrmanova}{A list of enrichment results for manova-derived gene sets.}
  \item{enrNetw}{A list containing enrichment results for network-derived gene sets, including differential enrichment.}
  \item{enr_manova.Netw}{A list of intersecting enrichment terms between manova and network results, for both same and opposite trends.}
}
}
\description{
This function performs a functional analysis combining manova pairwise results and network critical nodes. It generates a summary table, performs enrichment analysis for manova and network results, and identifies common enrichment terms.
}
\examples{
\dontrun{
# Example usage
names_of_groups <- c("Group1", "Group2")

fun_list <- c(betweenness, centroids, bridging_centrality)
preprocData <- preprocessing_data(dataset, names_of_groups,1, "TotSigNorm" )
manova_pairwise_results <- manova_pairwise(preprocData$data_norm,names_of_groups, 1 )
UNA <- unweighted_network_analysis(names_of_groups = names_of_groups,
                                   data_unique = preprocData$data_unique,
                                   fun_list = fun_list)
WNA <- weighted_netowrk_analysis(data_grouped_even_dim = data_grouped_even_dim,
                                 names_of_groups = names_of_groups,
                                 fun_list = fun_list)
unweighted_CN <- UNA$CriticalNodes
weighted_CN   <- WNA$CriticalNodes
names_of_groups <- c("Group1", "Group2")
tax_ID <- 9606
categories <- c("Process", "KEGG", "WikiPathways")

functional_analysis(manova_pairwise_results, unweighted_CN, weighted_CN,
                    names_of_groups, tax_ID, categories)

}
}
