% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{Normalization}
\alias{Normalization}
\title{Normalization}
\usage{
Normalization(dataset, normType)
}
\arguments{
\item{dataset}{The dataset that we want to normalize Protein x Samples}

\item{normType}{type of nornalization, choose from:
ln - natural logarithm transformation
Znorm - zeta normalization
MinMax - Min-Max scaling
Robust - Robust scaling
UnitVector - Unit vector scaling
TotSigNorm - total signal normalization
MaxSigNorm - maximum signal normalization
RowSigmaNorm - row sigma normalization}
}
\value{
data normalized
}
\description{
Normalization
}
