% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_mapping.R
\name{protein_mapping}
\alias{protein_mapping}
\title{Map Protein IDs Using StringDB}
\usage{
protein_mapping(proteins_ids, tax_ID, splitting_API)
}
\arguments{
\item{proteins_ids}{A character vector of protein IDs that need to be annotated.}

\item{tax_ID}{An integer specifying the taxonomy ID of the species to use for the annotation.}

\item{splitting_API}{An integer specifying the number of proteins to process per batch in the API call.}
}
\value{
A list with two elements:
\describe{
  \item{mapped_id}{A data frame with the mapped protein IDs.}
  \item{annotation.df}{A data frame containing the annotations for the proteins that map to more than one gene.}
}
}
\description{
This function maps a vector of protein IDs using the StringDB API.
It splits the input into smaller chunks to retrieve annotations and mapped IDs,
which are then combined into a final list of data frames.
}
\details{
The function interacts with the `rba_string_annotations` and `rba_string_map_ids` functions
from the `rbioapi` package to fetch annotations and map protein IDs. The input vector is split
into smaller chunks to avoid overloading the API. Annotations with more than one gene are kept
in the final output.
}
\examples{
\dontrun{
  proteins <- c("P12345", "Q67890", "A11111")
  mapped_id <- protein_mapping(proteins, tax_ID = 9606, splitting_API = 100)
  mapped_id  # Mapped IDs
}

}
\seealso{
\code{\link[rbioapi]{rba_string_map_ids}}
}
