% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_analysis.R
\name{quantitative_analysis}
\alias{quantitative_analysis}
\title{Quantitative Analysis of Dataset}
\usage{
quantitative_analysis(
  dataset,
  names_of_groups,
  gene_column,
  data_grouped_full = NULL,
  significance_manova,
  ...
)
}
\arguments{
\item{dataset}{A data frame containing the data to analyze.}

\item{names_of_groups}{A character vector specifying the names of the groups for analysis.}

\item{gene_column}{An integer or character specifying the column in the dataset containing gene names.}

\item{data_grouped_full}{List of data frames with the same number of rows.}

\item{significance_manova}{A numeric argument indicating the significance threshold for Multivariate Analysis of Variance}

\item{...}{Additional arguments passed to the underlying functions for manova, DAve, DCI, and fold change calculations.}
}
\value{
A list containing the results of the quantitative analysis:
\describe{
  \item{manova_results}{The results from manova analysis if more than two groups are present, with DAve and Fold Change indexes.}
  \item{manova_pairw_results}{The results from pairwise manova, with DAve and Fold Change indexes.}
  \item{mds_plot}{The results from Multi Dimensional Scaling if more than two groups are present.}
  \item{mds_plot_results}{The results from pairwise Multi dimensional Scaling.}
  \item{volcano_plots}{Volcano plot for each pairwise comparison.}
}
}
\description{
This function performs quantitative analysis on a dataset, including Multivariate Analysis of Variance (manova), pairwise manova (if more than two groups are provided), and calculation of additional indices such as DAve, DCI, and fold change.
}
\examples{
# Example dataset
data <- data.frame(
matrix(runif(1500),ncol = 15,
       dimnames = list(paste0("Gene_",1:100),
                       paste0("Group",rep(c(1,2,3),each = 5),"_",c(1:5))))
)
data$Gene <- paste0("Gene_",1:100)

# Perform quantitative analysis
qa <- quantitative_analysis(data, names_of_groups = c("Group1", "Group2", "Group3"),
                            gene_column = "Gene", significance_manova = 0.05)

}
