% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Summary Table for Network Analysis Results}
\usage{
summary_table(LDA_pairw.results, CN.unweighted, CN.weighted)
}
\arguments{
\item{LDA_pairw.results}{The results from pairwise LDA analysis.}

\item{CN.unweighted}{The list of critical nodes identified in unweighted networks.}

\item{CN.weighted}{The list of critical nodes identified in weighted networks.}
}
\value{
A list containing:
\describe{
  \item{summaryLDA}{A data frame summarizing the LDA analysis results.}
  \item{summaryUnweightedNetw}{A data frame summarizing the critical nodes identified in unweighted networks.}
  \item{summaryWeightedNetw}{A data frame summarizing the critical nodes identified in weighted networks.}
  \item{summary}{A merged data frame containing the combined summary table with the following columns:}
  \itemize{
    \item{GeneName:} The name of the gene.
    \item{Various metrics:} Metrics derived from LDA results, unweighted, and weighted networks.
    \item{rawSum:} The sum of non-missing values across all metrics.
    \item{combinedScore:} A score calculated based on different metrics with weighted contributions.
  }
}
}
\description{
This function generates a summary table that merges results from pairwise LDA, unweighted network critical nodes, and weighted network critical nodes. It calculates a combined score for each gene based on various metrics.
}
\examples{
# Example usage
\dontrun{
LDA_results <- list(gene_matrix = data.frame(GeneName = c("Gene1", "Gene2"),
                        LDA_Score = c(5, 3)))
CN_unweighted <- list(CN.M = data.frame(GeneName = c("Gene1", "Gene3"),
                          Specific_Centrality_unweighted = c(2, 4)))
CN_weighted <- list(CN.M = data.frame(GeneName = c("Gene1", "Gene2"),
                        Centrality_weighted = c(3, 6)))

summary_table(LDA_pairw.results = LDA_results, CN.unweighted = CN_unweighted,
                    CN.weighted = CN_weighted)
}
}
