% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{Transpose a Dataset}
\usage{
transpose(dataset, gene_column = 1)
}
\arguments{
\item{dataset}{A data frame to transpose.}

\item{gene_column}{The name or index of the column to use for setting the `colnames` of the transposed data. If set to 0, the `rownames` will be used as `colnames` Default is 1.}
}
\value{
A transposed data frame with updated column names.
}
\description{
This function transposes a given `data.frame`, optionally setting the column names of the transposed `data.frame` using a specified column from the original dataset or using the `rownames`.
}
\examples{
# Example with a simple dataset
data <- data.frame(
  Gene = c("Gene1", "Gene2", "Gene3"),
  Sample1 = c(1, 4, 7),
  Sample2 = c(2, 5, 8),
  Sample3 = c(3, 6, 9)
)

# Transpose using the "Gene" column as new column names
transpose(data, gene_column = "Gene")

# Transpose using row names as new column names
rownames(data) <- data$Gene
transpose(data, gene_column = 0)

}
