% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violin_betweennessRandom.R
\name{Violin_betweennessRandom}
\alias{Violin_betweennessRandom}
\title{Violin_betweennessRandom}
\usage{
Violin_betweennessRandom(graph, nIter = 100, Weights = NULL, set_seed = 1)
}
\arguments{
\item{graph}{A igraph object}

\item{nIter}{number of random iterations}

\item{Weights}{weights (as distances) to use computing the betweenness}

\item{set_seed}{random seed}
}
\value{
A ggplot object containing a distribution betweenness violin plot
}
\description{
Violin_betweennessRandom
}
