% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Violin_degreeRandom.R
\name{Violin_degreeRandom}
\alias{Violin_degreeRandom}
\title{Violin_degreeRandom}
\usage{
Violin_degreeRandom(graph, nIter = 100, set_seed = 1)
}
\arguments{
\item{graph}{A igraph object}

\item{nIter}{number of random iterations}

\item{set_seed}{random seed}
}
\value{
A ggplot object containing a distribution degree violin plot
}
\description{
Violin_degreeRandom
}
