% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_plot.R
\name{Volcano_Plot}
\alias{Volcano_Plot}
\title{Volcano Plot}
\usage{
Volcano_Plot(df, boundFC, thr_pvalue, Name_1 = "Group1", Name_2 = "Group2")
}
\arguments{
\item{df}{Data frame with columns named log2 (x axis), pv (y axis), gene (labels)}

\item{boundFC}{Numeric vector. Interval out of which labels are highlithed in red and labels are shown}

\item{thr_pvalue}{numeric value. Value under of which labels are highlithed in red and labels are shown}

\item{Name_1}{character Name of group 1}

\item{Name_2}{character Name of group 2}
}
\value{
A ggplot object
}
\description{
Volcano Plot
}
